Master

+++ src/main/java/cn/net/wangchenyu/finance/dao/FaultTypeDao.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by lenovo on 2016/7/17.
  */
-public interface FaultTypeDao extends CrudRepository<FaultType,Long> {
+public interface FaultTypeDao extends CrudRepository<FaultType,Integer> {
 }
